% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyGWR.R
\name{apply.GWR}
\alias{apply.GWR}
\title{Apply Geographically Weighted Regression (GWR)}
\usage{
apply.GWR(
  formula,
  data,
  longlat = TRUE,
  cl = NULL,
  fit.points = NULL,
  showProgressSteps = FALSE
)
}
\arguments{
\item{formula}{A formula object describing the model to be fitted.}

\item{data}{A spatial data frame containing the variables in the formula.}

\item{longlat}{Logical; if TRUE, coordinates are treated as longitude/latitude.}

\item{cl}{Optional cluster object for parallel processing.}

\item{fit.points}{Optional spatial data frame for making predictions.}

\item{showProgressSteps}{Logical; if TRUE, prints progress messages.}
}
\value{
A list containing the inputs and the GWR output.
}
\description{
This function applies Geographically Weighted Regression (GWR) to a given formula and dataset.
It allows for bandwidth selection and adaptive bandwidth, with optional progress messages.
}
\examples{
data(meuse.grid, package = "spEnviroDistr")
coordinates(meuse.grid) = ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
gridded(meuse.grid) <- TRUE

data(meuse, package = "spEnviroDistr")
coordinates(meuse) = ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")

cl <- createCluster(free = 2)
RESULT <- apply.GWR(sqrt(copper) ~ alt + dist, meuse, longlat = FALSE, cl = cl)

print(RESULT$output)

RESULT <- apply.GWR(log(copper) ~ alt + dist + elev, meuse, longlat = FALSE, cl = cl, fit.points = meuse.grid)
}
