% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Downscale Function for Spatial Data}
\usage{
downscale(response, pred.lr, pred.hr, method = "gwr", cl = NULL)
}
\arguments{
\item{response}{SpatRaster. The response variable raster (e.g., temperature) at low resolution.}

\item{pred.lr}{SpatRaster. The low-resolution predictor raster(s) (e.g., elevation at low resolution).}

\item{pred.hr}{SpatRaster. The high-resolution predictor raster(s) (e.g., elevation at high resolution) to be used for prediction.}

\item{method}{Character. The method used for downscaling. Options are "lm" for linear regression, "glm" for generalized linear model, and "gwr" for geographically weighted regression. Default is "gwr".}

\item{cl}{Optional. A cluster object for parallel processing. Default is NULL.}
}
\value{
A list containing two elements: the fitted model and the predicted high-resolution raster.
}
\description{
This function performs downscaling of spatial data using regression methods, including linear regression (lm), generalized linear model (glm), and geographically weighted regression (gwr).
It ensures that all input rasters have the same CRS and extent, resamples the response raster if necessary, and checks the overlap percentage between the high-resolution predictor and low-resolution predictor rasters.
}
\details{
The function checks that all rasters share the same CRS and extent. If the extents are not identical, it resamples the response raster to match the low-resolution predictor raster using bilinear interpolation.
It also calculates the overlap percentage between the high- and low-resolution rasters and stops execution if the overlap is less than 60\%.

The response raster is the variable of interest at a lower resolution (e.g., temperature), while \code{pred.lr} and \code{pred.hr} are the predictors at low and high resolutions, respectively (e.g., elevation). If the method is "lm" or "glm", it fits the model using the response and low-resolution predictor raster data and returns the model and the predicted high-resolution raster. If the method is "gwr", it performs geographically weighted regression using the provided spatial data and returns the result.
}
\examples{
library(terra)

# Load sample data
data(dem)
data(elev)
data(tavg)

# Convert to SpatRaster objects
tavg <- rast(tavg)   # Response raster (e.g., temperature)
elev <- rast(elev)   # Low-resolution predictor (e.g., elevation at low resolution)
dem <- rast(dem)     # High-resolution predictor (e.g., elevation at high resolution)
dem<-aggregate(dem,2)

# Apply the downscale function with GWR method
result <- downscale(response = tavg, pred.lr = elev, pred.hr = dem, method = "gwr")

# View the model and predicted high-resolution raster
print(result$model)
par(mfrow=c(2,1))
plot(tavg)
plot(result$result)

}
