% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillNA.R
\name{fillNA}
\alias{fillNA}
\title{Fill NA values in a Raster}
\usage{
fillNA(x, w, i, verbose = FALSE)
}
\arguments{
\item{x}{A \code{SpatRaster} object from the \code{terra} package.}

\item{w}{A matrix representing the focal window to be used. Typically, this is a square matrix with odd dimensions.}

\item{i}{An integer specifying the number of iterations to apply the focal filter.}

\item{verbose}{A logical value indicating whether to print the iteration number. Default is \code{FALSE}.}
}
\value{
A \code{SpatRaster} object with NA values filled.
}
\description{
This function fills NA values in a raster using a focal median filter.
}
\details{
The function iteratively applies a focal median filter to fill NA values in a raster. The filter is applied \code{i} times. If \code{verbose} is \code{TRUE}, the function prints the iteration number.
}
\examples{
# Load the meuse.grid data
data(meuse.grid, package = "spEnviroDistr")
coordinates(meuse.grid) = ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
gridded(meuse.grid) <- TRUE

# Convert meuse.grid to SpatRaster
r <- rast(meuse.grid)

# Define the focal window
w <- matrix(1, 3, 3)

# Fill NA values with 3 iterations and verbose output
result <- fillNA(r[["alt"]], w, 3, verbose = TRUE)

# Plot the result
plot(c(r[["alt"]],result))

}
