% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spData.R
\docType{data}
\name{spData}
\alias{spData}
\title{Average Air Temperature and Elevation Data for Cauquenes (WorldClim and ASTER)}
\format{
A list of three \code{SpatRaster} objects from the \code{terra} package:
\describe{
\item{tavg}{Annual Mean Temperature (BIO1) in °C, derived from WorldClim 2.1, representing the average air temperature for the period 1970-2000.}
\item{elev}{Elevation in meters, derived from WorldClim 2.1, representing the altitude above sea level.}
\item{dem}{Digital Elevation Model (DEM) from ASTER, representing the terrain elevation in meters with a spatial resolution of 30 meters.}
}
}
\source{
\itemize{
\item WorldClim 2.1: \url{https://www.worldclim.org/data/worldclim21.html}
\item NASA/METI/AIST/Japan Spacesystems, and U.S./Japan ASTER Science Team. ASTER Global Digital Elevation Model V003. 2018, distributed by NASA EOSDIS Land Processes DAAC, \url{https://doi.org/10.5067/ASTER/ASTGTM.003}
}
}
\usage{
data(dem, package = "spEnviroDistr")

data(elev, package = "spEnviroDistr")

data(tavg, package = "spEnviroDistr")
}
\description{
This dataset contains three rasters: Average Annual Air Temperature (tavg), Elevation (elev) from WorldClim 2.1, and a Digital Elevation Model (dem) from ASTER, all clipped to the extent of the Cauquenes region in Chile.
}
\details{
This dataset is ideal for spatial environmental analysis and ecological modeling. The three rasters provide valuable information about the climate and elevation of the Cauquenes region, which is located in central Chile. The data can be used for spatial interpolation, geostatistical analysis, or climate modeling.

The rasters are provided in the EPSG:4326 (WGS84) coordinate reference system.
}
\examples{
# Load the dataset
data(dem)
data(elev)
data(tavg)
plot(dem)
plot(elev)
plot(tavg)


}
\references{
Fick, S.E. and Hijmans, R.J., 2017. WorldClim 2: New 1 km spatial resolution climate surfaces for global land areas. \emph{International Journal of Climatology}, 37(12): 4302-4315. DOI: \doi{10.1002/joc.5086}

NASA/METI/AIST/Japan Spacesystems, and U.S./Japan ASTER Science Team. ASTER Global Digital Elevation Model V003. 2018, distributed by NASA EOSDIS Land Processes DAAC, \doi{10.5067/ASTER/ASTGTM.003}
}
\keyword{datasets}
