% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spEnviroDistr-package.R
\docType{package}
\name{spEnviroDistr-package}
\alias{spEnviroDistr}
\alias{spEnviroDistr-package}
\title{spEnviroDistr: Spatial Environmental Data Distribution Tools}
\description{
This package provides tools for handling, analyzing, and visualizing spatial environmental data. It includes datasets and functions specifically tailored for geostatistical analysis and environmental modeling. The package offers the 'meuse' and 'meuse.grid' datasets, which are well-known examples for demonstrating spatial data analysis techniques.
}
\author{
\strong{Maintainer}: Guillermo Fuentes-Jaque \email{g.fuentes@renare.uchile.cl} (\href{https://orcid.org/0000-0002-7864-4899}{ORCID})

}
\keyword{internal}
