% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPredictors.R
\name{toPredictors}
\alias{toPredictors}
\title{Convert a List of Layers into a RasterStack of Predictors}
\usage{
toPredictors(layerList, layerNames = NULL, reference = NULL)
}
\arguments{
\item{layerList}{A list of raster layers (\code{SpatRaster} objects) to be combined into a \code{SpatRaster} stack. These layers should represent different environmental variables or predictors.}

\item{layerNames}{An optional character vector specifying names for the layers in the resulting \code{SpatRaster}. If not provided, the function will attempt to derive names from the input list.}

\item{reference}{An optional reference layer (\code{SpatRaster}) to which all layers in \code{layerList} will be resampled. If not provided, the first layer in \code{layerList} will be used as the reference.}
}
\value{
A \code{SpatRaster} object where all layers have been resampled to match the reference layer and assigned appropriate names.
}
\description{
This function converts a list of raster layers into a \code{SpatRaster} object, aligning them to a common reference grid for spatial analysis.
It allows for resampling of the layers to match a specified reference layer and assigns appropriate names to the resulting raster stack.
}
\details{
The function first checks if \code{layerNames} are provided or if they can be derived from the \code{layerList}.
If the length of \code{layerNames} does not match the length of \code{layerList}, the function will stop and return an error.

If no reference layer is provided, the function will use the first layer in \code{layerList} as the reference for resampling.
Each layer in \code{layerList} is then resampled to match the reference layer's resolution and extent.
The resulting raster layers are combined into a \code{SpatRaster} stack, which can be used as input for further spatial analysis or modeling.
}
\note{
The function uses parallel processing to speed up the resampling process when dealing with large datasets.
}
\examples{
library(terra)

# Example raster layers
r1 <- rast(nrows=10, ncols=10)
r2 <- rast(nrows=10, ncols=10)
values(r1) <- runif(ncell(r1))
values(r2) <- runif(ncell(r2))

# Convert to list
layers <- list(r1, r2)

# Use the first layer as reference for resampling
predictors <- toPredictors(layers)

# Check the names and extent
names(predictors)
ext(predictors)

}
