% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSpatData.R
\name{toSpatData}
\alias{toSpatData}
\title{Convert SpatRaster to Spatial Points Data Frame}
\usage{
toSpatData(layers, na.rm = TRUE)
}
\arguments{
\item{layers}{A \code{SpatRaster} object representing the raster layers to be converted.}

\item{na.rm}{Logical. If \code{TRUE}, NA values are removed from the data (default is \code{TRUE}).}
}
\value{
A \code{SpatialPointsDataFrame} object containing the raster data with corresponding spatial coordinates.
}
\description{
This function converts a \code{SpatRaster} object into a \code{SpatialPointsDataFrame},
preserving the coordinate reference system (CRS) and including \code{x} and \code{y} coordinates.
It is useful when you need to work with raster data in a spatial format suitable for analysis
in various spatial data packages.
}
\details{
The function first converts the \code{SpatRaster} into a tibble with x and y coordinates using the \code{as_tibble} function.
The resulting tibble is then converted into a spatial object with spatial points (x and y coordinates).
The coordinate reference system (CRS) is assigned from the input raster using the \code{proj4string} function.
}
\note{
The function assumes that the \code{SpatRaster} object has a valid CRS.
If the CRS is missing or invalid, the resulting spatial object may not align properly with other spatial data.
}
\examples{
logo <- rast(system.file("ex/logo.tif", package="terra"))
names(logo) <- c("red", "green", "blue")
spat_data <- toSpatData(logo)
head(spat_data)
}
