% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCluster.R
\name{createCluster}
\alias{createCluster}
\title{Create a Cluster for Parallel Processing}
\usage{
createCluster(free = 1)
}
\arguments{
\item{free}{Integer. Number of cores to leave free. Default is 1.}
}
\value{
A cluster object.
}
\description{
This function detects the number of cores available on the machine,
subtracts the number of cores specified to be left free, and creates a cluster
using the remaining cores.
}
\details{
This function detects the total number of available cores on the machine, subtracts
the number of cores specified to be left free, and creates a cluster with the remaining cores.
On Linux systems, the cluster type is determined using the \code{snow} package's \code{getClusterOption()}.
On other operating systems, it creates a default cluster without specifying the type.

Note: This function requires the \code{parallel} and \code{snow} packages to be installed.
}
\examples{
\dontrun{
cl <- createCluster(free = 2)
}
}
