% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toGrid.R
\name{toGrid}
\alias{toGrid}
\title{Converts a Data Frame into a Spatial Grid Object}
\usage{
toGrid(x, coords, crs_str)
}
\arguments{
\item{x}{A data frame containing the spatial data. Must include columns for coordinates.}

\item{coords}{A character vector of length 2 specifying the names of the columns
in the data frame that represent the x and y coordinates.}

\item{crs_str}{A character string or a \code{CRS} object representing the coordinate
reference system to be assigned to the spatial object. If \code{NULL}, no CRS is assigned.}
}
\value{
A spatial grid object with the specified coordinates and CRS.
}
\description{
This function transforms a data frame into a spatial grid object, suitable for spatial analysis.
The function uses the columns specified in the \code{coords} parameter to define the spatial coordinates
in the data frame. The coordinate reference system (CRS) is specified by the \code{crs_str} parameter,
which defines how the spatial coordinates are interpreted. The result is a grid object that can
be used for spatial analysis in various spatial data packages and tools.
}
\details{
The function assumes that the data frame contains columns with coordinate data
specified by the \code{coords} parameter. It converts the data frame to a spatial grid object
using the \code{sp} package functions. The CRS is set if a valid CRS string or object is provided.
}
\examples{
# Load the meuse.grid data
data(meuse.grid, package = "spEnviroDistr")

# Convert meuse.grid to a spatial grid with specified coordinates and CRS
meuse.grid <- toGrid(meuse.grid, coords = c("x", "y"), crs_str = "+init=epsg:28992")

# Plot the spatial grid
spplot(meuse.grid, "alt")
spplot(meuse.grid, c("dist", "soil", "ffreq", "part.a", "part.b", "elev"))

}
